filmhub-backend/requirements.txt
django==4.2
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
djangorestframework-simplejwt
pytest
pytest-django
pytest-cov
